web-app/requirements.txt
-i https://pypi.org/simple
blinker==1.9.0; python_version >= '3.9'
click==8.3.1; python_version >= '3.10'
flask==3.1.2; python_version >= '3.9'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
markupsafe==3.0.3; python_version >= '3.9'
werkzeug==3.1.4; python_version >= '3.9'
dnspython==2.8.0; python_version >= '3.10'
pymongo==4.15.5; python_version >= '3.9'
pymongo>=4.7
python-dotenv
PyJWT
bcrypt
mongomock